chore(Tests): tidy up namespaces, usings and whitespace
