+++ src/Google.Maps.Test/Geocoding/GeocodingRequestTests.cs
@@ -1,12 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using NUnit.Framework;
-using Google.Maps.Geocoding;
-using System.Reflection;
 
-namespace Google.Maps.Test
+using NUnit.Framework;
+
+namespace Google.Maps.Geocoding
 {
 	[TestFixture]
 	public class GeocodingRequestTests
