+++ src/Google.Maps.Test/ViaLatLngTests.cs
@@ -1,11 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using NUnit.Framework;
-using Google.Maps;
 
-namespace Google.Maps.Test
+using NUnit.Framework;
+
+namespace Google.Maps
 {
 	[TestFixture]
 	public class ViaLatLngTests
@@ -95,6 +93,5 @@ namespace Google.Maps.Test
 			ViaLatLng ViaLatLng4 = new ViaLatLng(0d, 0d);
 			Assert.IsFalse(ViaLatLng1.Equals(ViaLatLng4));
 		}
-
 	}
 }
